+++ spec/methods/payment/payment.ts
@@ -13,7 +13,6 @@ export const paymentBasePath = `${apiBasePath}/${RootResources.payment}/stripe`;
  */
 export interface StripePlans {
   standard: number
-  ['standard-tiered']?: number
   professional: number
   enterprise: number
 }
@@ -35,8 +34,7 @@ export function isStripePlans(val:any): val is StripePlans {
  */
 export function trialStripePlan():StripePlans {
   return {
-    standard: 0,
-    ['standard-tiered']: 1,
+    standard: 1,
     professional: 0,
     enterprise: 0
   }